use a single MetricsModelMock definition
